test(app): isolate perf timeline budget contract
